requirements.txt
Flask==0.10.1
Flask-Cors==1.9.0
GitPython==0.1.7
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==1.6
Werkzeug==0.9.6
gunicorn==19.1.1
itsdangerous==0.24
peewee==2.3.3
redis==2.10.3
six==1.8.0
wsgiref==0.1.2
yolk==0.4.3
